app/requirements.txt
Flask>=2.0.0
rembg>=2.0.0
Pillow>=9.0.0
Werkzeug>=2.0.0